+++ scirust-ids/src/detectors/port_scan.rs
@@ -199,7 +199,7 @@ mod tests {
         let mut window = FlowWindow::new(0.0, 60.0);
         for port in 1..=50
         {
-            let mut f = Flow::new("10.0.0.1", "10.0.0.2", 40000 + port as u16, port);
+            let mut f = Flow::new("10.0.0.1", "10.0.0.2", 40000 + port, port);
             f.start_time = port as f64 * 0.5;
             f.end_time = f.start_time + 0.1;
             f.packets_out = 1;
