+++ scirust-ids/src/engine.rs
@@ -331,7 +331,7 @@ mod tests {
 
         for port in 1..=30
         {
-            let mut f = Flow::new("10.0.0.1", "10.0.0.2", 40000 + port as u16, port);
+            let mut f = Flow::new("10.0.0.1", "10.0.0.2", 40000 + port, port);
             f.start_time = port as f64 * 0.1;
             f.end_time = f.start_time + 0.05;
             f.packets_out = 1;
